Tests: Use pytest to parameterize tests of examples
The test of the examples are dynamically generated based on the contents
of the examples folder. Before this commit, they were all part of one
single test. It was therefore hard/impossible to parallelize (with e.g.
pytest-xdist), and it also meant the entire test would fail if a single
example had an error. Further inspection of the logs was necessary to
figure out which example(s) had failed.

Because we are mostly relying on pytest anyway during CI/CD, and
checking the exit code of the examples is more of a "nice to have"
compared to the other tests, we deviate from the norm of writing all
tests in unittest format. The benefits of parallelization, logging and
cleaner code outweigh the downsides of relying on a particular testing
package.