QTH/Buffer: Fix buffer equations
The previous formulation did not take into account the situation where
multiple buffers were presents. Indeed, the flows in the hot and cold
tank could be different in that scenario. We now fix this by linking the
volume in the cold tank to the cold pipe discharge.
To avoid overconstraining the problem, we add one additional constraint
for each extra buffer. Meaning that if the model contains k buffer, k-1
constraints are added by this commit.