Allow pipes in network to be disconnected
Thus far only pipes connected to buffers would be disconnectable. The
HeatMixin was not capable of making the dH-Q constraints trivial in case
the pipe was disconnected.

Note that a pipe disconnected symbol is _per timestep_. So it could be
connected on one timestep, but not on another. Generally this is not
desirable, as you lose out on the physics of heat losses.

Also note that the QTHMixin already did its own thing to disable
the head loss constraints per timestep. No changes are therefore needed
to the QTHMixin logic.

This commit serves as a step up to topology optimization. In that case,
if a pipe is disconnected, it will be for _all_ timesteps.