HeatMixin: Add support for pipe diameters size optimization
In the design of networks, it is often not yet known what size pipe need
to be put into the ground. The HeatMixin is, being MILP by design,
suitable for these discrete types of topology decisions. Furthermore, it
is able to do this while taking into account various other types of
constraints that need to be met, e.g. min/max pressure constraints and
velocity/flow limits.

A new `pipe_classes()` method is added, taking as input the pipe name.
This gives the user freedom to specify the possible pipe classes (of
instance `PipeClass`) for a certain pipe, and varying the available
options from pipe to pipe. A PipeClass with `.diameter == 0.0` is a
special case, meaning that one of the options is to _not_ lay the pipe.

The `pipe_diameter_symbol_name()` method can be used to retrieve the
diameter symbol name associated with a certain pipe for e.g.
minimization methods.

See the `pipe_diameter_sizing` example added in this commit to show how
this new functionality can be used.

Note that there are a few combinations of options not allowed yet when a
pipe has multiple pipe class options. All these can be technically
implemented if needed (at some cost to performance and implementation
complexity).

- Heat losses: should be zero for a pipe with multiple diameter options
- Minimum velocity: should be zero if any pipe has multiple diameter
  options
- Rate of temperature/discharge: should be infinite

User-facing exceptions and assertions are in place to make sure these
conditions are met.