HeadLossMixin._hn_pipe_head_loss: Allow floats input and match output type
Aside from the np.ndarray type, it can also be useful to calculate the
head loss for a single scalar discharge value. In that case, we should
make sure to also return a single scalar float. This is automatically
correct for the LINEAR and CQ2_* head loss options, but for the
DW-inequality we have to unpack.