HeatMixin: Fix handling of .NO_HEADLOSS in Big-M constraints
The `__get_maximum_total_head_loss` would fail if .NO_HEADLOSS was set,
as it called `_hn_pipe_head_loss()` for every pipe. The latter raises an
exception if it is called when .NO_HEADLOSS is set.

Because the "maximum total head loss" value does not make sense in a
situation where we do not handle head losses, we just return NaN. Note
that we also have to make sure to never use this NaN value, by skipping
some constraints in case .NO_HEADLOSS is set.