HeatMixin: Optionally neglect heat losses in pipes
Sometimes it can be useful to know what would happen if e.g. one changes
a pipe diameter, but keeps all things other than hydraulics equal. That
can help in understanding why and to what extent the heat losses in pipe
affect the solution. It can also be useful if the user does not care
about the heat losses, or does not know how to estimate the insulation
properties.