QTHMixin: Optionally allow variable temperature drop over demands
A new option `demand_temperature_option` is added that controls what
the temperature drop and/or the return temperature of a demand is. The
default behavior is still to fix the temperature drop at the demands
(`FIXED_DT`). In some cases a more realistic approach is to set a
minimum return temperature, accompanied by an upper bound on the
temperature drop(`MIN_RETURN_MAX_DT`).

In case the user wants to specify a more intricate relationship, e.g.
one where the return temperature drops sublinearly with the supply
temperature, they can use the `FREE` option. They are then free to add
their own constraints/relationships.