HeatMixin: Fix Big-M of head loss - dH constraints
For the inequality head loss options, we relate a head loss symbol to
the discharge. This head loss symbol is always positive, whereas .dH can
be both positive and negative (depending on flow direction). The Big-M
however was not big enough to let `head_loss` take on values up to
`max_head_loss`.

For example, let's say that `head_loss = 1.0`, and that the flow
direction is negative (`flow_dir = 0`). The second inequality constraint
states that:

  dh - head_loss + flow_dir * big_m >= 0

which works out to:

  dh - 1.0 >= 0

  dh >= 1.0

If we then take the first inequality constraint, we get:

  -dh - head_loss + (1 - flow_dir) * big_m >= 0

  -dh - 1.0 + big_m >= 0

  1.0 - 1.0 + big_m >= 0

  big_m >= 2.0

Up until this commit, we used a Big-M of `max_head_loss`. In other
words, we never allowed the head loss of a pipe to approach half its
maximum. The proper Big-M needs to be at least _twice_ the maximum head
loss, which is the bug that this commit fixes. Because we do not want
our Big-Ms to be overly tight, we use an additional safety factor of 2.

Note that for the .LINEAR head loss option the constraints were fine
as-is. That is because for that option, we relate the dH directly to Q,
and let the `head_loss` symbol be free.