HeatMixin: Add support for heat losses in topology optimizaton
The previous commit (02bfaddc) added support for topology optimization
of pipe diameters. It required the heat losses of pipes with multiple
diameter options to be zero. This commit gets rid of that particular
limitation.

Note that this commit makes it possible for the user to specify pipe
class options sharing a diameter, but having different U-values/heat
losses.