HeatMixin: Optionally force heat loss zero if pipe is disconnected
We add a new option `heat_loss_disconnected_pipe` to force the heat loss
and heat flow rate of a pipe to zero if it is disconnected. By default
(option is `True`), a pipe has heat loss even if it is disconnected, as
it would still contain relatively hot water in reality. We also do not
want minimization of heat production to lead to overly disconnecting
pipes.

In some scenarios it is hydraulically impossible to supply heat to these
disconnected pipes (Q is forced to zero), in which case this new option
can be set to `False`.