HeadLossMixin: Fix Linearized Darcy-Weisbach for disconnectable pipes
When passing a CasADi MX symbol for `is_disconnected` to the head loss
method, the shape matches that of the other ca.MX symbols, e.g. (N, 1)
for N timesteps. We then also need to be sure to broadcast the
disconnected MX symbol just like we do the discharge and head loss when
LINEARIZED_DW is used. If we do not, CasADi will error out with a shape
mismatch for the constraint expression.