Tests/simple_buffer: Fix initial heat to buffer in history test case
A heat flow rate of 1000 W corresponds to a discharge of 8E-6 m3/s, and
a velocity of 0.4 mm/s. The minimum velocity we impose is 5 mm/s, which
means that the pipe _has_ to be disconnected at t0. So we end up with a
situation where Q is zero, but the heat flow rate is not.

Although interesting, the test case was not written to cover that
scenario. We want to enforce the initial condition to some that can be
attained, to check whether HeatMixin does not accidentally ignore it. We
fix the test case to describe a feasible situation. A heat flow rate of
12000 W corresponds to a velocity of 5.4 mm/s, so the pipe can be
connected at t0.