ESDL: Improve default maximum temperature if none specified
When converting an ESDL model to a QTH model, a maximum temperature is
needed to bound the temperature in the assets (e.g. to support goals on
them).

In case _no_ information is available, we increase the default from
110 C to 140 C. This is to support high temperature (water based) heat
networks that can go up to 120 C. In case any information is available
in the ESDL file, we can also choose a better (=lower) upper bound by
adding 10 C to the maximum temperature we find.

Note that this value is still just a default in case no maxTemperature
is specified for the respective asset. Also note that the user is free
to override this default value by overriding `esdl_qth_model_options()`.