ESDL: Bump ESDL version
We could not parse recently generated ESDL models due to the addition of
the "related" attribute to pipes. This commit updates the ESDL
library/pyecore files such that we can parse such ESDL files again.

The "related" attribute is used to relate hot to cold pipes. Note that
we however still rely on the naming to pair up pipes.