QTH/Buffer: Fix initial conditions as bounds
As in HeatMixin, we now set all the initial conditions of a buffer
(volume and temperature of the hot and cold tanks) by overwriting the
bounds of the corresponding variables. This is both for clarity reason,
since now all the initial conditions are set within the same method, as
well as improved robustness.