Heat: Improve overall scaling of heat equations
We move a factor of 100.0 of the heat to the state entry, to reduce the
coefficient in front of the heat variables. This particularly helps the
scaling/range of the constraints that relate the heat loss (if it is
variable/optional) to the heat in- and out of a component.