HeatMixin: Reduce Big-M of pipe dH-Q disconnectable
The total heat loss in the system is likely already a large overestimate
of the maximum head loss over a certain pipe. Instead of using a factor
of 2, it therefore makes sense to lower this to improve the overall
equation scaling.

Note that in exceptional cases the Big-M can still be rather tight, e.g.
a network consisting of a single pipe. Those cases are however so small
that they hardly lead to solver issues.