HeatMixin: Convert assertion into logging of warning/error
We do a check in post() to see if the flow velocity / discharge matches
the flow direction boolean. Especially for CBC, the assertion could be
triggered readily with very small exceedences (e.g. < 0.1 mm/s). Instead
of this assertion causing the run to fail, we should log a warning or
error depending on the severity of the exceedence.

Note that we also convert the check from a relative one (based on the
nominal of the discharge) to an absolute one (based on the velocity).
This is because the minimum velocity is specified in absolute terms as
well (default 0.5 mm/s), and scaling in general should be such that the
nominal velocity is approximately 1 m/s.