ESDLMixin: Change _RunInfoReader to not parse ESDL model
The RunInfoReader class should concern itself with reading the RunInfo
file. It should not _also_ parse the ESDL model. This way the
RunInfoReader class can be easily reused by others that are also
interested in reading the contents of a RunInfo.xml, possibly before
ESDLMixin gets to its __init__ routine.