HeatMixin: Add method that exposes optimized PipeClass
This commit adds a method `get_optimized_pipe_class` that returns the
PipeClass that was selected for a certain pipe. Note that this method
can be called after every priority (`priority_completed()`).

For pipes that do not have any PipeClass options (i.e. when
`pipe_classes(pipe)` returns an empty list), a KeyError is raised when
`get_optimized_pipe_class` is called.