ESDL: Set heat demand upper bound to inf as default
If the user does not set a value, then this is read as a 0.0. This means
that the user is forced to specify a maximum value for each heat
demand for the mpc to be able to have a meaningful run. Note that this
change does not influence any part of the code nor its
current applications (i.e., warmingup mpc).