Style: Fix new flake8-bugbear errors (B020)
With 22.3.20, flake8-bugbear started to report a style error "found for
loop that reassigns the iterable it is iterating with each iterable
value". It probably thinks there might be an issue in iterating over
self.<something> with a variable self.<something else>.

To resolve the issue, we note that self.__n_times is not actually used
outside of the affected "optimize()" method, so we make it a local
variable.