ESDL: Warn if a buffer has both volume and capacity set
Volume and capacity both describe the size of a buffer. If both are
specified we continue execution by using the volume, instead of erroring
out. We will however warn the user that we are ignoring their specified
capacity.

Also slightly refactor the branching code into a if-elif, which makes it
clear that volume has the higher priority.