HeatMixin: Fix disconnected pipe assertion accuracy
We forgot to divide the variable we are checking by its nominal, to get
to the value (of the decision variable) that the solver is dealing with.
With heat in the order of 1e4 - 1e5, we were placing an unrealistic
expectation on accuracy of the output of the solver.

This commit also removes the relative tolerance part of the comparison,
as the target value that we compare (and want) to is zero.