Heatpump functionality with connection to e-grid is added. This is first functionality where a  multi commodity interaction is tested. It is now possible to have a heatpump that has a connection to both the electricity grid and the heat network. Heatpumps with solely a connection to the heat network are still allowed.
Fixes are done to the electricity grid, to allow power dissipation while maintaining minimal voltage requirements and ensuring maximum current requirements. Also scaling of the heatpump and electricity grid have been improved.

Furthermore general improvement in heat_mixing for heat-exchangers and heatpumps where the minimal power is removed. An tolerance is added to ensure that the asset is disabled when it produces zero heat.