This commit adds producer profile constraints using a scaled profile with its maximum capacity. The producer must produce less or equal to that profile, will not necessarily exactly follow profile. This functionality can for instance be used in sizing problems with aqua or solar thermal sources.
The buffer previously heat could not flow through the buffer to compensate losses in the return network. This has been fixed by simplifying the logic for extracting heat from the network similar to the demand. 

The heat loss symbol for the return pipes are now set separately. Issues concerning the bounds were fixed, and issues for setting the heat loss symbol in the pipe diameter sizing problem.

Some basic tests on the matching of heat demands, energy conservation and flow balance are added. They are used in the producer_profile tests and should be included in the other tests as well.