23 change to absolute heat (#42)
This commit is a very large one effecting the whole structure of the optimization problem. In this commit we change from a relative thermal power description to an absolute thermal power variable. This is done to support networks with different supply and return configurations, it can also be used for temperature cascading in a single hydraulically coupled network and thus giving more design flexibility.

The main is in all heat to discharge constraints, we now use an equality constraint on the outgoing flow for every asset. Meaning that we equate Q * rho * cp * T == Heat for outgoing flows, and inequalities for the heat carried in the pipes. This means that the heat can decrease in the network to compensate losses, but that the losses and thus flow will always be over-estimated with the temperature for which no temperature drops are modelled.  

A new test-case has been added that used temperature cascading in a single hydraulically coupled network to test the new heat to discharge constraints. An example running a municipality network has been added but is not part yet of the ci/cd pipeline as we need to be able to access a influxdb database.

This new heat to discharge formulation also means that a minimum velocity in the pipe is needed to ensure that no artificial heat is created in the inequality constraints of the pipes. This minimum velocity is in the order of 1.e-3-1.e-4 m/s, thus this has no major effect on the applicability of the real life result.

To enhance the computational speed of the problem two measures were taken. One, if a supply and return pipe are there neighbors we only create a single variable for their optimization in the problem to minimize the solution space. For our ESDL interface we assume that the neighbors have a "_ret" in the return network pipe. Secondly, we incorporated a two stage problem, where we in the first stage optimize without heat-losses, and in a second stage optimize with heat losses but constraining the solution space of the pipe sizes with the pipe sizes found in stage one and allowing for one DN size larger. As the problem without heat-losses is much much faster we can achieve much faster computational times. The run_municipality example should now take around 10-15 minutes on a normal laptop with highs.

For the ordering variables of the pipe classes we do have to create separate heat loss ordering variables for the return and supply network as the ordering might be different for the hot and cold side depending on the exact u_1 and u_2 values.

Another effect is that the carriers do not have a dedicated supply or return label anymore. Instead they just have a temperature now, thus the naming of the variables is also changed.

---------

Co-authored-by: Rojer <jim.rojer@tno.nl>