Start for creation HT-ATES models with varying temperature state, Created the continuous and discrete variable for ATES. Added constraint to have same discrete ATES temperature as the chosen network temperature during discharging.
Still requires;
constraints to link continuous and discrete temperature ATES,
constraints to determine temperature changes of continous temperature, both of heatloss (down) and when loading (up).
constraints to determine the efficiency/heatloss of the ates.
constraints in asset_sizing_mixin to change the maximum Q based on the aggregation count optimum.