52 modelling temperature state in ates and constraining network temperature (#105)
Created HT-ATES models with varying temperature state,
Created the continuous and discrete temperature variables for ATES.
Added constraint to have same discrete ATES temperature as the chosen network temperature during discharging and network temperature should be equal or larger during charging. Temperature losses are created when discharging the ATES based on linearised inequalities from convex functions. Full explanation of philosophy is in the code.
To utilise this functionality, the option: include_ates_temperature_options needs to set to True. Default value is False as it requires specific setup of the problem.

Added limitation of the flowrate for sizing with aggregation count for ATES and also included limitation on heat discharge of ATES based on ATES temperature and the maximum flowrate.

Additional cyclic constraint on the temperature is required in the workflow when ATES temperatures are used.
Additional constraint in the workflow is required to ensure that only HEX or HP is active at the same moment in time, in the future this might be covered by compound assets.


bugfixes pipe heat to discharge constraints for varying temperature when option no_heat_losses is selected.

Inequality on continues and discrete temperature to ensure we can specifically select the boolean with the closest related variable, this does result in a slightly slower computation, but it ensures that in other configurations also the correct boolean is chosen. This allows to have large temperature steps even at smaller timesteps and thus solving relatively fast and will also allow to work with grow workflow with variable timestep.
In the future maybe look at a two-stage approach where first large timesteps are used as a hot-start or further constrained problem for the second stage with smaller timesteps or ordering of temperature booleans over time. Furthermore the problem is very challenging for open-source solvers.

---------

Co-authored-by: Rojer <jim.rojer@tno.nl>
Co-authored-by: Thijs van der Klauw <52158067+Tklauw-TNO@users.noreply.github.com>
Co-authored-by: Thijs van der Klauw <thijs.vanderklauw@tno.nl>